requirements.txt
fastapi
uvicorn
pypdf
faiss-cpu
numpy
sentence-transformers
python-multipart

